Modularización para mejor entendimiento de #7, #8, #9, #10 y #11
