+++ internal/palet.go
@@ -1,0 +1,8 @@
+package internal
+
+// Palet representa un tipo de palet con una cantidad específica
+type Palet struct {
+	Tipo     string
+	Cantidad int
+	Estado   string // Estado del palet
+}
